fix: intl build: add rest notfound page, move tsx to locale
