+++ src/app/[locale]/metadata.ts
